separate unpacks and inline them
this is the bulk of the manual work, which is
moving away from unpack(). to do that, we split a
single unpack call into unpacks for every field,
which will later be replaced by int() and sint()
(and sometimes bytes()).

in some places I went ahead and used int() directly