introduce primitive system, error check system
introduce in_object() and in_list() / in_list_item()
to prepare for producing JSON output, and use it
in the core parsing.

also introduce handle_errors(), which catches errors
and prints a hexdump of the full parser's contents
when that happens. use it in core parsing.

indentation is now increased in in_object(), not in
subparser().

now that we have a dedicated function for error
handling, we can use it in main() to catch errors
in the root context too! behavior should otherwise
be the same (TODO: test)