introduce field(describe=...)
this parameter to field specifies a function to
compute the textual description of the value,
and is shown in parenthesis next to the formatted
value (if descriptions are enabled).

the pattern is common enough.

an important FIXME is that the describe function
is only called when descriptions are enabled (vs the
current idiom which always computed the description)
which may be a problem for functions that perform assertions