✨ senc box parsing
the senc box is particularly tricky because in order to
parse it, you need to derive the value of the
PerSampleIVSize variable for the sample group you are in,
which is NOT EASY as it can vary per group and per fragment.

for now, implement the parsing but require the user to
manually specify the size, which should cover most use cases.