Do not consider data handlers for stsd parsing
last_handler_seen should not be set after parsing data handlers. When
we're parsing a video track, the stsd should be parsed like a video
track, as an alias data handler or what not.

This change makes the stsd of the QuickTime files actually show
contents. They contain errors because there are a bunch of
QuickTime-specific fields that have not been implemented -- that's
coming in a following patch.